+++ js/data.js
@@ -1,0 +1,21 @@
+const DESCRIPTIONS = ['Bonjour!', 'Это я в отпуске.', 'Это мой кот. Он красивый.', 'Привет. Как дела?', 'Я и закат.', 'Очень усердно пишу код.', 'На даче с бабушкой. Набрал уже 10кг.'];
+
+const COMMENTS = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const NAMES = [
+  'Кекс',
+  'Феликс',
+  'Бетховен',
+  'Лисса',
+  'Барсик',
+  'Марсик',
+  'Маруся',
+  'Кейси',
+  'Персиваль',
+  'Муся',
+  'Лейси',
+  'Фенек',
+  'Юта'
+];
+
+export { DESCRIPTIONS, COMMENTS, NAMES };
